Move ebreak* logic from take_trap into instructions. (#1006)
Now that logic only affects ebreak instructions, and does not affect
triggers that also cause a trap to be taken.

Fixes #725. Although like Paul, I don't have a test for this case.

Introduce trap_debug_mode so so ebreak instructions can force entry into
debug mode.

Note: This commit comes from the spike upstream. When rebasing cosim we
can ditch this commit though note it's subtly different as we've
modified ebreak to not report a PC value when it's an exception.