Allow hardware triggers to go off without using the mmu tlb
For our cosimulation env, the Spike mmu TLB functionality is not used.
Hardware triggers are only checked on instuction fetch if there is an
entry in the mmu TLB, so it would never trigger with our implementation.
This commit removes this check (of tlb_insn_tag).

Also, move the checking code to after we have fetched all the bytes of
an instruction. Currently, as spike fetches instructions in 2-byte accesses,
if the match address is set within an instruction, it is possible for the
trigger to go off. E.g. if you have valid, uncompressed RV32 instructions at 0x4
and 0x8, and the trigger set to 0x6, with the previous version this trigger
would go off.

TODO/N.B. Discard this commit when we rebase onto latest upstream.
https://github.com/lowRISC/riscv-isa-sim/pull/21#pullrequestreview-1147256925